Fix biome gen weight

+++ src/main/java/cavern/data/Miner.java
@@ -283,7 +283,7 @@ public class Miner implements IMiner
 
 			tag.setString("Name", record.getKey().getBlockName());
 			tag.setInteger("Meta", record.getKey().getMeta());
-			tag.setInteger("Count", record.getValue().intValue());
+			tag.setInteger("Count", record.getValue());
 
 			list.appendTag(tag);
 		}
@@ -334,7 +334,7 @@ public class Miner implements IMiner
 	{
 		Integer ret = MINING_POINTS.get(block, meta);
 
-		return ret == null ? 0 : ret.intValue();
+		return ret == null ? 0 : ret;
 	}
 
 	public static int getPointAmount(IBlockState state)
